design header

+++ app/controllers/texts_controller.rb
@@ -1,5 +1,6 @@
 class TextsController < ApplicationController
   before_action :user_params
+  before_action :texts, except: :index
 
   def index
     @texts = Text.all
@@ -11,7 +12,7 @@ class TextsController < ApplicationController
 
   def create
     Text.create(text_params)
-    redirect_to texts_path
+    redirect_to @user 
   end
 
   def edit
@@ -20,21 +21,27 @@ class TextsController < ApplicationController
 
   def update
     text = Text.find(params[:id])
-    text.update(text_params)
-    redirect_to user = User.find(1)
+    text.update(text_params) if text.user.id == current_user.id
+    redirect_to @user
   end
 
   def destroy
-    text = Text.find(2)
-    text.destroy
+    text = Text.find(params[:id])
+    text.destroy if text.user.id == current_user.id
+    redirect_to @user
   end
 
   private
+
   def text_params
-    params.require(:text).permit(:tag, :text)
+    params.require(:text).permit(:tag, :text).merge(user_id: current_user.id)
   end
 
   def user_params
     @user = User.find(1)
+  end
+
+  def texts
+    @texts = Text.all
   end
 end
